:hammer: Refactor: code structure
